requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
